Enable connection to Firebase database and saving of names
Authentication and the database are two distinct things, so to store user-related data besides what is related to authentication, a separate database is used. This database is essentially a JSON tree. Rules have been set up so only authenticated users can read and write data to the nodes corresponding to their accounts UID (which comes from authentication). The first and last name in sign up are now saved to the databse, and the main activity has a section to save random str to the database with respect to the current user as a test